+++ ibm-mq/src/main/scala/jms4s/ibmmq/ibmMQ.scala
@@ -6,7 +6,7 @@ import cats.implicits._
 import com.ibm.mq.jms.MQConnectionFactory
 import com.ibm.msg.client.wmq.common.CommonConstants
 import io.chrisdavenport.log4cats.Logger
-import jms4s.jms.{ JmsConnection, JmsContext }
+import jms4s.jms.JmsContext
 
 object ibmMQ {
 
@@ -24,40 +24,6 @@ object ibmMQ {
   case class QueueManager(value: String) extends AnyVal
   case class Channel(value: String)      extends AnyVal
   case class ClientId(value: String)     extends AnyVal
-
-  def makeConnection[F[_]: Sync: Logger: ContextShift](
-    config: Config,
-    blocker: Blocker
-  ): Resource[F, JmsConnection[F]] =
-    for {
-      connection <- Resource.make(
-                     Logger[F].info(s"Opening Connection to MQ at ${hosts(config.endpoints)}...") >>
-                       blocker.delay {
-                         val connectionFactory: MQConnectionFactory = new MQConnectionFactory()
-                         connectionFactory.setTransportType(CommonConstants.WMQ_CM_CLIENT)
-                         connectionFactory.setQueueManager(config.qm.value)
-                         connectionFactory.setConnectionNameList(hosts(config.endpoints))
-                         connectionFactory.setChannel(config.channel.value)
-                         connectionFactory.setClientID(config.clientId.value)
-
-                         val connection = config.username.map { (username) =>
-                           connectionFactory.createConnection(
-                             username.value,
-                             config.password.map(_.value).getOrElse("")
-                           )
-                         }.getOrElse(connectionFactory.createConnection)
-
-                         connection.start()
-                         connection
-                       }
-                   )(
-                     c =>
-                       Logger[F].info(s"Closing Connection $c at ${hosts(config.endpoints)}...") *>
-                         blocker.delay(c.close()) *>
-                         Logger[F].info(s"Closed Connection $c.")
-                   )
-      _ <- Resource.liftF(Logger[F].info(s"Opened Connection $connection at ${hosts(config.endpoints)}."))
-    } yield new JmsConnection[F](connection, blocker)
 
   def makeContext[F[_]: Sync: Logger: ContextShift](
     config: Config,
