+++ tests/src/test/scala/jms4s/basespec/providers/ActiveMQArtemisBaseSpec.scala
@@ -1,19 +1,20 @@
 package jms4s.basespec.providers
 
 import cats.data.NonEmptyList
-import cats.effect.{Blocker, ContextShift, IO, Resource}
+import cats.effect.{ Blocker, ContextShift, IO, Resource }
 import jms4s.activemq.activeMQ
-import jms4s.activemq.activeMQ.{ClientId, Config, Endpoint, Password, Username}
+import jms4s.activemq.activeMQ._
 import jms4s.basespec.Jms4sBaseSpec
-import jms4s.jms.{JmsConnection, JmsContext}
+import jms4s.jms.JmsContext
 
 trait ActiveMQArtemisBaseSpec extends Jms4sBaseSpec {
-  override def connectionRes(implicit cs: ContextShift[IO]): Resource[IO, JmsConnection[IO]] =
+
+  override def contextRes(implicit cs: ContextShift[IO]): Resource[IO, JmsContext[IO]] =
     Blocker
       .apply[IO]
       .flatMap(
         blocker =>
-          activeMQ.makeConnection[IO](
+          activeMQ.makeContext[IO](
             Config(
               endpoints = NonEmptyList.one(Endpoint("localhost", 61616)),
               username = Some(Username("admin")),
@@ -23,20 +24,5 @@ trait ActiveMQArtemisBaseSpec extends Jms4sBaseSpec {
             blocker
           )
       )
-
-  override def contextRes(implicit cs: ContextShift[IO]): Resource[IO, JmsContext[IO]] = Blocker
-    .apply[IO]
-    .flatMap(
-      blocker =>
-        activeMQ.makeContext[IO](
-          Config(
-            endpoints = NonEmptyList.one(Endpoint("localhost", 61616)),
-            username = Some(Username("admin")),
-            password = Some(Password("passw0rd")),
-            clientId = ClientId("jms-specs")
-          ),
-          blocker
-        )
-    )
 
 }
